Initiate major redesign: make this a proper Reader add-on
The text layer has turned out to be inadequate for a good reading
experience due to many reasons:

    * Reliability issues: the layer sometimes contains inacurrate text.
    * Lost visual cues: no bold text, italics, etc. + single color text.
    * Lost semanitcs (esp. on font resize): no headings, sub-heads, etc.
    * Changing font is anyway buggy without text reflow.

On the whole it was a suboptimal solution to go customizing a component
made for enabling text-copying, never meant for visual presentation. It
was easy, and I was tempted by quick results. The correct solution is to
modify the canvas-drawing process of PDF.js, which I had not considered.
This commit initiates the transition in that direction.

The focus is now only on color. It might be possible to tweak PDF.js's
font-rendering process also, but that is surely a lot more complicated,
I am afraid to even try. Modifying fonts was never the main intent any-
way; it just came along with the text-layer solution (and finally led to
the Terminal analogy).

As a first step, this commit strips the canvas-hiding code and replaces
the 'Terminal-mode' references.