Implement contrast correction and image blending
We need to maintain the legibility of content when making color changes.
Ensuring that the WCAG-recommended minimum contrast ratios are applied
seems to be the sensible thing to do. The contrast ratio is calculated
w.r.t. the immediate canvas background.

Inverting images can aid in their legibility e.g. a transparent PNG with
black text would otherwise be unreadable on a dark background. Blending
is useful for the opposite effect e.g. making a b/w figure merge nicely
with a light background.