Add a power button to the toolbar, update icons
Since the button on the viewer's toolbar will now be used to show/hide
the add-on's UI, we need another button to toggle the reader on/off. Now
that's one additional click/tap/keypress, but I could not conceive of a
better design. It is commonplace, should be a familiar to users.

A keyboard shortcut could be added for quick toggle; or a gesture for
touch devices (e.g. long press on the toolbar button, akin to Safari on
iOS). Deferred to future.