Make last used settings persist across sessions
All settings are saved and applied automatically upon launch *except*
the reader mode itself. We do not want to jump at the user any time they
open a PDF; they might have forgotten that reader was on, or they might
be using it only occasionally.

Providing an option to configure this intentionally can be considered.