Store separate prefs for light/dark system themes
This lets users always have a reader theme matching the OS theme. A new
`options' interface has been provided to [en|dis]able this feature.

Also, with the redesigned UI, reader mode turns on automatically up on
launch if last used. As commented in a20ac3e, an option to configure
this is desirable, and `options' is apt for this purpose also.

Host deployments can directly set `doq.options' in localStorage for now
to configure both these options.