Major optimization: save canvas only when needed
We don't need to save canvas ImageData after each shape is drawn. Since
the calculation of only text-styles depend on the canvas state, we need
to save it only after the _last_ shape in a sequence is drawn, i.e. just
before a text sequence starts. So, we in fact have to _forget_ the data
while shapes are being drawn.

Since getImageData is a heavy operation, the performance benefit of this
approach is significant, especially while rendering PDFs having lots of
diagrams/graphs/other shapes.