Drop storing canvas, analyze textBg only if needed
We need to ensure legibility only in the cases of either the text or its
background being an accent color; if both are greyscale, the legibility
present in original colors should not be affected by the interpolation.
This also helps us narrow down the need to access canvas imageData, that
we can do away with caching it altogether, vastly reducing memory usage.

A minor trade-off is that we can no longer reliably wrap strokeText; it
might get called to paint exactly on top of a painted fillText, and then
the imageData at the center might not represent the actual background.

This and the preceding commit converge on one goal, which is to optimize
rendering of majorly black-on-white PDFs (which is our main target) with
the default settings. So far, much of the heavylifting was unnecessarily
being done to be able to handle a few edge cases, at the cost of memory
and performance.