Enable non-viewer integration and general usage
This commit adds a JS API layer to directly control the addon core, i.e.
the theme engine, without the UI. This allows doq to be used in projects
that use only the PDF.js core library, and not the viewer.

In fact, doq can now be used by any Canvas-based document-rendering lib,
via the API, to add theme support. The PDF.js viewer add-on can serve as
a reference implementation for such usage.