[Annotations] Add support for the Highlight editor
Highlight annotations (added in PDF.js 4.1) need a different treatment
because while the editors are drawn on the annotation editor layer, the
actual highlights are implemented as SVGs _inside_ the .canvasWrapper.
But the mappings between the two are behind an obscure #private class
field, and directly modifying the annotation objects introduces some
overheads (like keeping track of the actual color) and ugly workarounds.

As with FreeText annotations, it is better to modify the SVGs directly in
reader mode via CSS/styles, and a MutationObserver turns out to be the most
flexible choice for covering all case scenarios.