[Engine] Force use of software canvas in Firefox
This is a late follow up to 89a6f95, further addressing the regression
introduced by 1c3ecda. Turns out getImageData() gets painfully slow when
the canvas is hardware accelarated (which is now Firefox default on Mac,
Linux and Android). While Chrome has always warned about this, it still
rendered fast, and strangely, disabling HW acceleration makes it worse.

The solution is far from ideal, but rather preferable to the alternative
of reverting 1c3ecda and going back to caching canvasData. Whether it is
too drastic needs to be seen, or we need to find a more innovative fix.