[Options] Allow users to force software rendering
This commit introduces a new option `softwareRender` to allow users to
manually configure disabling GPU instead of guessing the browser. This
cleans up the temporary fix introduced in the previous commit, which we
have decided to keep as the PDF.js team has addressed the issue and has
chosen to do the same via a preference (mozilla/pdf.js#18238 on GitHub).