[Engine] bring back canvas caching, but improved
We have come a full circle since 1c3ecda. This commit partially reverts
those changes, because our patchy solution to the hardware acceleration
conundrum (8ec9869) has turned out to be... well, totally unreliable.

We reintroduce canvas caching in an improved avatar, similar to what we
did before, but with these important changes:

  * keep the pixel data of only the currently rendered page, identified
    by the context; pages are rendered one at a time, so it is safe to
    discard the previous canvases when a new one starts rendering.
  * store the colors extracted from canvas data in a text-color map per
    page; as the document is fixed, text is always rendered in the same
    order, so we need not redo the same extraction every time.

Pages are identified by a special data attribute on the canvas that the
user can set. This is to leave the engine code generic and independent.

These changes reduce both processing and memory footprint of the cache.
We have retained other changes from and since 1c3ecda, like acting only
on pages with backgrounds and/or accents, updating the cache only when
needed, etc. Also, we are able to rewrap `strokeText` as a side effect.