[Annots] Rewrite and refactor annotation handling
This is a major overhaul of both SVG and non-SVG annotation handling.
The patch achieves the following things:

  * fully seperate add-on logic from core by moving the remaining viewer
    related code from `lib/` to `app/`, completing the work of 7b0c8de;
  * generalise SVG annotation handling and correctly recolor ink (draw)
    annotations in PDF.js 4.9+, as they have now been made SVG as well;
  * better reassign various parts of annotations event handling to the
    different modules; and
  * rename, reorder and clean up the `annots.js` functions and exports.

This refactor should also finally enable *doq* lib users to theme PDF.js
annotations in non-viewer environments, as they can pass the appropriate
container elements to the exported functions of `annots.js`.