[Core] Simplify image blending in drawImage handler
Instead of drawing the mask on the canvas and waiting for PDF.js, we can
directly blend the image with the mask and return the result for drawing
with the normal source-over. This is logically simpler and less prone to
breaking with complicated multi-image arrangements (sometimes found for
title decorations in books). This also calls for some apt renamings.